+++ NEWS.md
@@ -1,5 +1,8 @@
 # bigrquery (development version)
 
+* `bq_field()` can now pass `description` parameter which will be applied
+  in `bq_table_create()` call (@byapparov, #272).
+
 # bigrquery 1.1.0
 
 ## Improved type support